-Hierarchy restructure -General refactoring

+++ Firebrick4/Firebrick/Random.cpp
@@ -1,7 +1,7 @@
 #include "pch.h"
 #include "Random.h"
 
-int CRandom::Random(const float n)
+int CRandom::Random(const int n)
 {
 	return static_cast<int>(static_cast<double> (rand()) / (RAND_MAX + 1) * (n + 1));
 }
